requirements.txt
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
azure-ai-vision-imageanalysis==1.0.0
azure-core==1.33.0
certifi==2025.1.31
charset-normalizer==3.4.1
defusedxml==0.7.1
distro==1.9.0
filelock==3.18.0
fsspec==2025.3.2
h11==0.16.0
h5py==3.13.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
isodate==0.7.2
Jinja2==3.1.6
jiter==0.10.0
langdetect==1.0.9
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
onnx==1.17.0
openai==1.79.0
opencv-python==4.11.0.86
packaging==24.2
pillow==11.2.1
protobuf==6.30.2
pyclipper==1.3.0.post6
pydantic==2.11.4
pydantic_core==2.33.2
pypdfium2==4.30.1
python-doctr==0.11.0
python-dotenv==1.1.0
PyYAML==6.0.2
RapidFuzz==3.13.0
requests==2.32.3
scipy==1.15.2
setuptools==75.8.0
shapely==2.1.0
six==1.17.0
sniffio==1.3.1
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uv==0.6.14
wheel==0.45.1
